refacto: create service SetResultFrontIntoArray
